Refactor: PostApiControllerSliceTest와 PostControllerSliceTest를 controller 패키지로 이동
