+++ src/test/java/kr/co/pinup/posts/controller/PostApiControllerSliceTest.java
@@ -1,4 +1,4 @@
-package kr.co.pinup.posts;
+package kr.co.pinup.posts.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.FilterChain;
@@ -6,7 +6,6 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.ConstraintViolationException;
-import kr.co.pinup.comments.model.dto.CommentResponse;
 import kr.co.pinup.comments.repository.CommentRepository;
 import kr.co.pinup.comments.service.CommentService;
 import kr.co.pinup.members.Member;
@@ -16,18 +15,15 @@ import kr.co.pinup.members.model.enums.MemberRole;
 import kr.co.pinup.members.repository.MemberRepository;
 import kr.co.pinup.members.service.MemberService;
 import kr.co.pinup.oauth.OAuthProvider;
-import kr.co.pinup.oauth.OAuthService;
-import kr.co.pinup.postImages.model.dto.CreatePostImageRequest;
 import kr.co.pinup.postImages.model.dto.PostImageResponse;
 import kr.co.pinup.postImages.service.PostImageService;
-import kr.co.pinup.posts.controller.PostApiController;
+import kr.co.pinup.posts.Post;
 import kr.co.pinup.posts.exception.post.PostNotFoundException;
 import kr.co.pinup.posts.model.dto.CreatePostRequest;
 import kr.co.pinup.posts.model.dto.PostResponse;
 import kr.co.pinup.posts.model.dto.UpdatePostRequest;
 import kr.co.pinup.posts.repository.PostRepository;
 import kr.co.pinup.posts.service.PostService;
-import kr.co.pinup.security.SecurityUtil;
 import kr.co.pinup.security.filter.AccessTokenValidationFilter;
 import kr.co.pinup.security.filter.SessionExpirationFilter;
 import kr.co.pinup.stores.Store;
